make `RemoteImageLoader` conform to `FeedImageLoader` and moved it to a separate file
